% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{alpha_to_logbf}
\alias{alpha_to_logbf}
\title{alpha_to_logbf}
\usage{
alpha_to_logbf(alpha, pi)
}
\arguments{
\item{alpha}{an L by p+1 matrix of posterior inclusion probabilites}

\item{pi}{per snp prior probability of causality}
}
\value{
L by p+1 matrix of log BF
}
\description{
convert alpha matrix to log BF matrix
}
\author{
Chris Wallace
}
