% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claudia.R
\name{approx.bf.p}
\alias{approx.bf.p}
\title{Internal function, approx.bf.p}
\usage{
approx.bf.p(p, f, type, N, s, suffix = NULL)
}
\arguments{
\item{p}{p value}

\item{f}{MAF}

\item{type}{"quant" or "cc"}

\item{N}{sample size}

\item{s}{proportion of samples that are cases, ignored if type=="quant"}

\item{suffix}{suffix to append to column names of returned data.frame}
}
\value{
data.frame containing lABF and intermediate calculations
}
\description{
Internal function, approx.bf.p
}
\details{
Calculate approximate Bayes Factors
}
\author{
Claudia Giambartolomei, Chris Wallace
}
