% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claudia.R
\name{approx.bf.z}
\alias{approx.bf.z}
\title{Internal function, approx.bf.z}
\usage{
approx.bf.z(zs, f, type, N, s, suffix = NULL)
}
\arguments{
\item{f}{MAF}

\item{type}{"quant" or "cc"}

\item{N}{sample size}

\item{s}{proportion of samples that are cases, ignored if type=="quant"}

\item{suffix}{suffix to append to column names of returned data.frame}

\item{z}{normal deviate associated with regression coefficient and its variance}
}
\value{
data.frame containing lABF and intermediate calculations
}
\description{
Internal function, approx.bf.z
}
\details{
Calculate approximate Bayes Factors using z-scores only, estimating variance as in approx.bf.p
}
\author{
Vincent Plagnol, Chris Wallace, Jamie Inshaw
}
