% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloccc.R
\name{approxv2.multinom}
\alias{approxv2.multinom}
\title{vcov estimation for H4, multinomial approx}
\usage{
approxv2.multinom(b1, b2, n00, n1, n2, f0)
}
\value{
list with diagnonal and off diagonal elements of vcov matrix
}
\description{
variance-covariance matrix for multinomial regression at a single SNP from coefficients from marginal logistic regressions
}
\details{
assume allele freq known in controls
}
\author{
Chris Wallace
}
