% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloccc.R
\name{bf2}
\alias{bf2}
\title{recreate bivariate logistic regression from univariate summary stats}
\usage{
bf2(b1, b2, v1, v2, rho, w1 = 0.04, w2 = 0.04)
}
\arguments{
\item{b1}{beta_1}

\item{b2}{beta_2}

\item{v1}{v(beta_1)}

\item{v2}{v(beta_2)}

\item{rho}{cor(snp1,snp2)}

\item{w1}{prior variance on b1 (beta_1 ~ N(b1,v); b1 ~ N(0,w1))}

\item{w2}{prior variance on b2}
}
\value{
log ABF of data under alternative vs null (w1=w2=0)
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Chris Wallace
}
