% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{bin2lin}
\alias{bin2lin}
\title{binomial to linear regression conversion}
\usage{
bin2lin(D)
}
\arguments{
\item{D}{dataset, as described in coloc.abf}
}
\value{
D, with original beta and varbeta in beta.bin, varbeta.bin, and beta and varbeta updated to linear estimates
}
\description{
Convert binomial to linear regression
}
\details{
Estimate beta and varbeta if a linear regression had been run on a
binary outcome, given log OR and their variance + MAF in controls

sets beta = cov(x,y)/var(x)
varbeta = (var(y)/var(x) - cov(x,y)^2/var(x)^2)/N
}
\author{
Chris Wallace
}
