% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloccc.R
\name{bstar_from_b}
\alias{bstar_from_b}
\title{recreate bivariate logistic regression from univariate summary stats}
\usage{
bstar_from_b(b1, b2, v1, v2, rho)
}
\arguments{
\item{b1}{beta_1}

\item{b2}{beta_2}

\item{v1}{v(beta_1)}

\item{v2}{v(beta_2)}

\item{rho}{cor(snp1,snp2)}
}
\value{
list of b1, b2 in bivariate regression
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Chris Wallace
}
