% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{bstar_from_bhat}
\alias{bstar_from_bhat}
\title{Estimate beta-star}
\usage{
bstar_from_bhat(bA, bB, fA, fB, rho)
}
\arguments{
\item{bA}{log OR for SNP A}

\item{bB}{log OR for SNP B}

\item{fA}{MAF of SNP A}

\item{fB}{MAF of SNP B}

\item{rho}{correlation between SNPs A, B (between -1 and 1)}
}
\value{
vector of length 2, bstar
}
\description{
Estimate the coefficients of the joint model (bstar) from those of the marginal model (bhat)
}
\details{
The marginal model fits

Y ~ A

Y ~ B

and returns separate estimates bhat for A and B.  This is the
usual practice in GWAS.  We need bstar, the estimates from the joint
model

Y ~ A + B
}
\author{
Stasia Grinberg
}
