% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.alignment}
\alias{check.alignment}
\title{check alignment}
\usage{
check.alignment(D, thr = 0.2)
}
\arguments{
\item{D}{a coloc dataset}

\item{thr}{plot SNP pairs in absolute LD > thr}
}
\value{
a plot for a visual check that alleles in your data are aligned the same way for the beta vector and the LD matrix
}
\description{
check alignment between beta and LD
}
\author{
Chris Wallace
}
