% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claudia.R
\name{check.dataset}
\alias{check.dataset}
\title{check.dataset}
\usage{
check.dataset(d, suffix = "")
}
\description{
Check dataset inputs for errors
}
\details{
Will call stop() unless a series of expectations on dataset input format are met
}
\author{
Chris Wallace
}
