% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_alignment}
\alias{check_alignment}
\alias{check.alignment}
\title{check alignment}
\usage{
check_alignment(D, thr = 0.2, do_plot = TRUE)

check.alignment(...)
}
\arguments{
\item{D}{a coloc dataset}

\item{thr}{plot SNP pairs in absolute LD > thr}

\item{do_plot}{if TRUE (default) plot the diagnostic}

\item{...}{arguments passed to check_alignment()}
}
\value{
proportion of pairs that are positive
}
\description{
check alignment between beta and LD
}
\author{
Chris Wallace
}
