\name{checklikechi}
\alias{checklikechi}
\title{Checking equivalence of test statistics}
\usage{
  checklikechi(V)
}
\arguments{
  \item{V}{Variance-covariance matrix}
}
\value{
  a number giving the maximum difference between the two
  statistics
}
\description{
  Code to determine whether, for a specific
  variance-covariance matrix for shared controls the form
  of the test statistic is the same under the two
  approaches
}
\author{
  Mary Fortune
}

