% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloccc.R
\name{coloc.cc}
\alias{coloc.cc}
\title{coloc.abf for two case control studies and (partially) shared controls}
\usage{
coloc.cc(dataset1, dataset2, n1 = args$ncse, n2 = args$ncse, n00 = (1 -
  dataset1$s) * dataset1$N, n01 = 0, n02 = 0, MAF, LD, p1 = 1e-04,
  p2 = 1e-04, p12 = 1e-05, method = c("multinom", "corr"))
}
\arguments{
\item{dataset1}{a list with the following elements
\describe{

\item{beta}{regression coefficient for each SNP from dataset 1}

\item{varbeta}{variance of beta}

\item{snp}{a character vector of snp ids. It will be used to merge dataset1 and dataset2.}

}}

\item{dataset2}{as above, for dataset 2}

\item{n1}{number of cases for dataset 1}

\item{n2}{number of cases for dataset 2}

\item{n00}{number of shared controls}

\item{n01}{number of controls unique to dataset 1}

\item{n02}{number of controls unique to dataset 2}

\item{MAF}{minor allele frequency of the variants, named vector}

\item{LD}{matrix giving correlation between SNPs.  NB - this is r, not rsquared!  Expect a mix of positive and negative values.  This should be estimated from reference data.}

\item{p1}{prior probability a SNP is associated with trait 1,
default 1e-4}

\item{p2}{prior probability a SNP is associated with trait 2,
default 1e-4}

\item{p12}{prior probability a SNP is associated with both traits,
default 1e-5}

\item{method}{"multonom" or "corr", depending on which
approximation (multinomial or correlation) you wish to use.}
}
\value{
a vector giving the number of SNPs analysed, and the
    posterior probabilities of H0 (no causal variant), H1 (causal
    variant for trait 1 only), H2 (causal variant for trait 2
    only), H3 (two distinct causal variants) and H4 (one common
    causal variant)
}
\description{
Bayesian colocalisation analysis for case control studies with (partially) shared controls
}
\details{
This function calculates posterior probabilities of different
causal variant configurations under the assumption of a single
causal variant for each trait.

Unlike coloc.abf, this specifically requires coefficients and
their standard errors in the case of partially shared controls.
}
\author{
Chris Wallace
}
