% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{coloc.process}
\alias{coloc.process}
\title{Post process coloc output}
\usage{
coloc.process(result, hits1 = NULL, hits2 = NULL, LD = NULL,
  r2thr = 0.01, p1 = 1e-04, p2 = 1e-04, p12 = 1e-06, LD1 = LD,
  LD2 = LD)
}
\arguments{
\item{result}{output from coloc.detail()}

\item{hits1}{list of independent signals from find.indep.signals() for trait 1}

\item{hits2}{list of independent signals from find.indep.signals() for trait 2}

\item{LD}{reference LD matrix}

\item{r2thr}{r2 threshold below which SNPs will be considered independent}

\item{p1}{prior probability a SNP is associated with trait 1, default 1e-4}

\item{p2}{prior probability a SNP is associated with trait 2, default 1e-4}

\item{p12}{prior probability a SNP is associated with both traits, default 1e-5}

\item{LD1}{reference LD matrix for dataset1, that will override LD if specified}

\item{LD2}{reference LD matrix for dataset2, that will override LD if specified}
}
\value{
a data.table of potentially multiple coloc analyses, one for each pair of hits1 and hits2 combinations
}
\description{
Post process a coloc.details result using masking
}
\author{
Chris Wallace
}
