% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocqq.R
\name{coloc.qq}
\alias{coloc.qq}
\title{Bayesian colocalisation analysis for two quantitative traits, measured on the same individuals}
\usage{
coloc.qq(N, VY1, VY2, corY, MAF, LD, p1 = 1e-04, p2 = 1e-04, p12 = 1e-05,
  W1 = 0.15^2 * VY1, W2 = 0.15^2 * VY2)
}
\arguments{
\item{N}{number of samples}

\item{VY1}{variance of trait 1}

\item{VY2}{variance of trait 2}

\item{corY}{cor(Y1,Y2)}

\item{MAF}{minor allele frequency of the variants, named vector}

\item{LD}{matrix giving correlation between SNPs.  NB - this is r,
not rsquared!  Expect a mix of positive and negative values.
This should be estimated from reference data.}

\item{p1}{prior probability a SNP is associated with trait 1,
default 1e-4}

\item{p2}{prior probability a SNP is associated with trait 2,
default 1e-4}

\item{p12}{prior probability a SNP is associated with both traits,
default 1e-5}

\item{W1}{the effect size of any variant for trait 1 under the
assumption it is causal is assumed to follow a normal with
mean 0 and variance W1}

\item{W2}{as above, for trait 2}
}
\value{
a list, with main entry "results" a vector giving the
    number of SNPs analysed, and the posterior probabilities of H0
    (no causal variant), H1 (causal variant for trait 1 only), H2
    (causal variant for trait 2 only), H3 (two distinct causal
    variants) and H4 (one common causal variant)
}
\description{
This function calculates posterior probabilities of different
causal variant configurations under the assumption of a single
causal variant for each trait.
}
\details{
Unlike coloc.abf, this specifically requires coefficients and
their standard errors - ie p values and MAF are not enough,
because the direction of effect matters
}
\author{
Chris Wallace
}
