% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{coloc.susie}
\alias{coloc.susie}
\title{run coloc using susie to detect separate signals}
\usage{
coloc.susie(dataset1, dataset2, susie.args = list(), ...)
}
\arguments{
\item{dataset1}{\emph{either} a coloc-style input dataset (see
\link{check.dataset}), or the result of running \link{runsusie} on such a
dataset}

\item{dataset2}{\emph{either} a coloc-style input dataset (see
\link{check.dataset}), or the result of running \link{runsusie} on such a
dataset}

\item{susie.args}{a named list of additional arguments to be passed to
\link{runsusie}}

\item{...}{other arguments passed to \link{coloc.bf_bf}, in particular prior
values for causal association with one trait (p1, p2) or both (p12)}
}
\value{
coloc.signals style result
}
\description{
colocalisation with multiple causal variants via SuSiE
}
\author{
Chris Wallace
}
