% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{coloc.susie_bf}
\alias{coloc.susie_bf}
\title{run coloc using susie to detect separate signals}
\usage{
coloc.susie_bf(dataset1, bf2, ...)
}
\arguments{
\item{dataset1}{a list with specifically named elements defining the dataset
to be analysed. See \code{\link{check.dataset}} for details.}

\item{bf2}{named vector of BF, names are snp ids and will be matched to column names of susie object's alpha}

\item{...}{arguments passed to \link{coloc.bf_bf}}
}
\value{
coloc.signals style result
}
\description{
coloc for susie output + a separate BF matrix
}
\author{
Chris Wallace
}
