% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas.R
\name{coloc.twas}
\alias{coloc.twas}
\title{Wrapper to use colocalization testing within a Bayesian model averaging
structure.}
\usage{
coloc.twas(df1, df2, snps = intersect(setdiff(colnames(df1), c(response1,
  stratum1)), setdiff(colnames(df2), c(response2, stratum2))),
  response1 = "Y", response2 = "Y", stratum1 = NULL, stratum2 = NULL,
  r2.window = 200, bayes = !is.null(bayes.factor), thr = 0.01,
  nsnps = 2, n.approx = 1001, bayes.factor = NULL, plot.coeff = FALSE,
  r2.trim = 0.95, quiet = FALSE, bma = FALSE, ...)
}
\arguments{
\item{df1, df2}{Each is a dataframe, containing response and
potential explanatory variables for two independent datasets.}

\item{snps}{The SNPs to consider as potential explanatory variables}

\item{response1, response2}{The names of the response variables in
\code{df1} and \code{df2} respectively}

\item{stratum1}{optional column name of df1 that gives stratum information}

\item{stratum2}{optional column name of df2 that gives stratum information}

\item{bayes}{Logical, indicating whether to perform Bayesian
inference for the coefficient of proportionality, eta.  If
\code{bayes.factor} is supplied, Bayes factors are additionally
computed for the specificed values.  This can add a little time as
it requires numerical integration, so can be set to FALSE to save
time in simulations, for example.}

\item{thr}{posterior probability threshold used to trim SNP list.
Only SNPs with a marginal posterior probability of inclusion
greater than this with one or other trait will be included in
the full BMA analysis}

\item{nsnps}{number of SNPs required to model both traits.  The BMA
analysis will average over all possible \code{nsnp} SNP models,
subject to \code{thr} above.}

\item{n.approx}{number of values at which to numerically
approximate the posterior}

\item{bayes.factor}{if true, compare specific models}

\item{plot.coeff}{deprecated}

\item{r2.trim}{for pairs SNPs with r2>\code{r2.trim}, only one SNP
will be retained.  This avoids numerical instability problems
caused by including two highly correlated SNPs in the model.}

\item{quiet}{suppress messages about how the model spaced is
trimmed for BMA}

\item{bma}{if true (default), average over models}

\item{...}{other parameters passed to \code{coloc.test}}

\item{family1, family2}{the error family for use in \code{glm}}
}
\value{
a \code{coloc} or \code{colocBayes} object
}
\description{
Performs the colocalisation tests described in Plagnol et al (2009)
and Wallace et al (2012), adapted to the specific situation of
twas, where we are comparing eQTL (trait 1) and GWAS (trait 2)
data, and where the number of SNPs may be very large.
}
\details{
The adaptations are first to split the SNPs according to LD, then
take through to colocalisation testing only those with some
evidence for association to the eQTL.  This is done by summing the
posterior support for H1, H3 and H4 given by the coloc.abf function
within each SNP group and using only SNPs in groups with posterior support of at least 0.05.

Note that SNPs groups that show association only to the GWAS trait
are ignored, because it is possible for GWAS variants acting
through different pathways to be physically proximal, and our focus
is on whether the eQTL can explain the GWAS trait.  Note that there
*may* exist GWAS variants that are proximal in evolutionary
distance (ie in some LD) that act through different pathways.  In
that case, this test should reject colocalisation.  This is not
appropriate, but developing a test that copes with this better will
require further work.

The second adaptation is to generate faster (though less accurate)
approximations to the Bayes Factors for multi SNP models that are
used to weight the colocalistaion test result for each pair of SNPs
by using the approximation that

ABF = ( BIC1 - BIC0 ) / 2

This is for reasons of speed, as more SNPs may enter this test.
}
\examples{

 ## simulate covariate matrix (X) and continuous response vector (Y)
 ## for two populations/triats Y1 and Y2 depend equally on f1 and f2
 ## within each population, although their distributions differ between
 ## populations.  They are compatible with a null hypothesis that they
 ## share a common causal variant
set.seed(1)
 X1 <- matrix(rbinom(2000,1,0.4),ncol=4)
 Y1 <- rnorm(500,rowSums(X1[,1:2]),2)
 X2 <- matrix(rbinom(2000,1,0.6),ncol=4)
 Y2 <- rnorm(500,rowSums(X2[,1:2]),5)
 
 boxplot(list(Y1,Y2),names=c("Y1","Y2"))
 
 ## fit and store linear model objects
 colnames(X1) <- colnames(X2) <- sprintf("f\%s",1:ncol(X1))
 summary(lm1 <- lm(Y1~f1+f2+f3+f4,data=as.data.frame(X1)))
 summary(lm2 <- lm(Y2~f1+f2+f3+f4,data=as.data.frame(X2)))
 

## test colocalisation using bma
df1=as.data.frame(cbind(Y1=Y1,X1))
df2=as.data.frame(cbind(Y2=Y2,X2))

result <- coloc.bma( df1, df2, snps=colnames(X1), response1="Y1", response2="Y2",
family1="gaussian", family2="gaussian",
nsnps=2,bayes.factor=c(1,2,3))
result
plot(result)

## test colocalisation when one dataset contains a stratifying factor in column named "s"
df1$s <- rbinom(500,1,0.5)
result <- coloc.bma( df1, df2, snps=colnames(X1), response1="Y1", response2="Y2",
stratum1="s",
family1="gaussian", family2="gaussian",
nsnps=2,bayes.factor=c(1,2,3))
result
plot(result)

}
\references{
Wallace et al (2012).  Statistical colocalisation of
    monocyte gene expression and genetic risk variants for type 1
    diabetes.  Hum Mol Genet 21:2815-2824.
    \url{http://europepmc.org/abstract/MED/22403184}

Plagnol et al (2009).  Statistical independence of the colocalized
association signals for type 1 diabetes and RPS26 gene expression on
chromosome 12q13. Biostatistics 10:327-34.
\url{http://www.ncbi.nlm.nih.gov/pubmed/19039033}
}
\author{
Chris Wallace
}
