% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coloc_test_data}
\alias{coloc_test_data}
\title{Simulated data to use in testing and vignettes in the coloc package}
\format{
A four of two coloc-style datasets. Elements D1 and D2 have a single
shared causal variant, and 50 SNPs. Elements D3 and D4 have 100 SNPs, one
shared causal variant, and one variant unique to D3. Use these as examples
of what a coloc-style dataset for a quantitative trait should look like.
}
\usage{
data(coloc_test_data)
}
\description{
Simulated data to use in testing and vignettes in the coloc package
}
\examples{
data(coloc_test_data)
names(coloc_test_data)
str(coloc_test_data$D1)
check_dataset(coloc_test_data$D1) # should return NULL if data structure is ok
}
\keyword{datasets}
