% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{colocABF-class}
\alias{colocABF-class}
\title{Class \code{"colocABF"} holds objects returned by the \code{coloc.abf} function}
\description{
Objects can be created by calls to the
function \code{\link{coloc.abf}()}.
}
\examples{
showClass("colocABF")
}
\seealso{
\code{\link{coloc.abf}}
}
\author{
Chris Wallace.
}
\keyword{classes}
