% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{colocDataset-class}
\alias{colocDataset-class}
\title{colocDataset class}
\arguments{
\item{object}{a list, containing all the data needed to describe one dataset}
}
\value{
object of class colocDataset
}
\description{
Class \code{"colocDataset"} holds the list of items in a dataset used by coloc.signals, coloc.susie
}
\author{
Chris Wallace
}
