% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claudia.R
\name{combine.abf}
\alias{combine.abf}
\title{combine.abf}
\usage{
combine.abf(l1, l2, p1, p2, p12)
}
\arguments{
\item{l1}{merged.df$lABF.df1}

\item{l2}{merged.df$lABF.df2}

\item{p1}{prior probability a SNP is associated with trait 1, default 1e-4}

\item{p2}{prior probability a SNP is associated with trait 2, default 1e-4}

\item{p12}{prior probability a SNP is associated with both traits, default 1e-5}
}
\value{
named numeric vector of posterior probabilities
}
\description{
Internal function, calculate posterior probabilities for configurations, given logABFs for each SNP and prior probs
}
\author{
Claudia Giambartolomei, Chris Wallace
}
