% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_datasets}
\alias{compare_datasets}
\title{Compare two datasets before running coloc}
\usage{
compare_datasets(dataset1, dataset2, thr = 0.8)
}
\arguments{
\item{dataset1}{first coloc dataset}

\item{dataset2}{second coloc dataset}

\item{thr}{minimum fine mapping posterior captured in by snps in
the other dataset}
}
\value{
a vector giving the fraction of the fine mapping posterior
in dataset1 captured by snps in dataset2, and the fraction of
the fine mapping posterior in dataset2 captured by snps in
dataset1
}
\description{
Datasets are merged before running coloc on the basis of snp
id. If a snp exists in one dataset and not the other, it will not
contribute to the coloc analysis. This may not be a problem if it
is a snp without much association signal. But if there is a
substantial change that one or more missing snps may be causal for
one dataset, this can cause misleading coloc
results. compare_datasets will allow you to check how much of the
fine mapping posterior in dataset1 is captured by snps in dataset2
and vice versa. Ideally the result should be two numbers close to
1, but the function will warn if either drops below thr.
}
\details{
Note, thr wasn't chosen on any sensible basis, just
intuition. This is a function to give the user information to help
decide whether to trust their coloc analysis or not, and different
values of thr may apply in different situations. Please take some
time to think about the right value for your analysis.
}
\author{
Chris Wallace
}
