% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varp.R
\name{eqtl_prior}
\alias{eqtl_prior}
\title{estimate prior probability of snp causality for gene expression based on
distance from TSS}
\usage{
eqtl_prior(dist)
}
\arguments{
\item{dist}{distance in bp from TSS for an possible eqtl}
}
\description{
estimate prior probability of snp causality for gene expression based on
distance from TSS
}
\examples{
dist=seq(0,1e+6,by=1000)
p=eqtl_prior(dist)
plot(dist,p,type="l")
}
\author{
Chrs Wallace
}
