% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eqtlgen_density_data}
\alias{eqtlgen_density_data}
\title{eQTLGen estimated distance density}
\format{
A data.frame with two columns, "rel_dist" the relative distance
to the TSS and "dens_value" the estimated value of the density. This
density is esimated from the eQTLGen data. For details of how the
density was estiamted see Pullin and Wallace (2024+).
}
\usage{
data(eqtlgen_density_data)
}
\description{
eQTLGen estimated distance density
}
\examples{
data(eqtlgen_density_data)
names(eqtlgen_density_data)
}
\keyword{datasets}
