% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{estgeno.2.ctl}
\alias{estgeno.2.ctl}
\alias{estgeno.2.cse}
\title{estgeno2}
\usage{
estgeno.2.ctl(fA, fB, rho)

estgeno.2.cse(coefs, G0)
}
\arguments{
\item{fA}{MAF of SNP A}

\item{fB}{MAF of SNP B}

\item{rho}{correlation between SNPs A, B (between -1 and 1)}

\item{coefs}{vector of marginal log OR for SNP A and SNP B}

\item{G0}{matrix of relative frequencies in controls, often derived from estgeno.2.ctl}
}
\value{
matrix giving relative frequency of each genotype. Element (i,j) gives frequency of A=i-1, B=j-1.
}
\description{
Estimate two snp frequency distibutions
}
\examples{
## two unlinked SNPs
estgeno.2.ctl(fA=0.5,fB=0.1,rho=0)
## two perfectly correlated SNPs
estgeno.2.ctl(fA=0.5,fB=0.5,rho=1)
## two unlinked SNPs, only one of which (A) is associated
G0=estgeno.2.ctl(fA=0.5,fB=0.5,rho=0)
estgeno.2.cse(coefs=c(log(2),log(1)), G0=G0)
}
\seealso{
estgeno1
}
\author{
Chris Wallace
}
