% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{find.best.signal}
\alias{find.best.signal}
\title{Pick out snp with most extreme Z score}
\usage{
find.best.signal(D)
}
\arguments{
\item{D}{standard format coloc dataset}
}
\value{
z at most significant snp, named by that snp id
}
\description{
Internal helper function
}
\author{
Chris Wallace
}
