% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{findends}
\alias{findends}
\title{trim a dataset to central peak(s)}
\usage{
findends(d, maxz = 4, maxr2 = 0.1, do.plot = FALSE)
}
\arguments{
\item{d}{a coloc dataset}

\item{maxz}{keep all snps between the leftmost and rightmost snp with |z| >
maxz}

\item{maxr2}{expand window to keep all snps between snps with r2 > maxr2
with the left/rightmost snps defined by the maxz threshold}

\item{do.plot}{if TRUE, plot dataset + boundaries}
}
\value{
logical vector of length d$position indicating which snps to keep
}
\description{
tries to be smart about detecting the interesting subregion to finemap/coloc.
}
\seealso{
findpeaks
}
\author{
Chris Wallace
}
