% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{finemap.bf}
\alias{finemap.bf}
\title{Finemap data through Bayes factors}
\usage{
finemap.bf(bf1, p1 = 1e-04)
}
\arguments{
\item{bf1}{named vector of BF, or matrix of BF with colnames (cols=snps, rows=signals)}
}
\value{
finemap.signals style result
}
\description{
Finemap one dataset represented by Bayes factors
}
\details{
This is the workhorse behind many finemap functions
}
\author{
Chris Wallace
}
