% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{finemap.indep.signals}
\alias{finemap.indep.signals}
\title{Finemap independent signals}
\usage{
finemap.indep.signals(D, LD = D$LD, method = c("mask", "cond"),
  r2thr = 0.01, sigsnps = NULL, pthr = 1e-06, zthr = qnorm(pthr/2,
  lower.tail = FALSE), maxhits = 10)
}
\arguments{
\item{D}{list of summary stats for a single disease, as defined
for coloc.abf}

\item{LD}{matrix of signed r values (not rsq!) giving correlation
between SNPs}

\item{method}{identify secondary signals by masking primary signals or conditioning on them. Conditioning requires that the effect estimates are allelically aligned with the data used to generate the reference LD matrix}

\item{r2thr}{if mask==TRUE, all snps will be masked with r2 >
r2thr with any sigsnps. Otherwise ignored}

\item{sigsnps}{SNPs already deemed significant, to condition on}

\item{pthr}{when p > pthr, stop successive conditioning}

\item{zthr}{alternative to specifying pthr, stop when abs(z) <
zthr}

\item{maxhits}{maximum depth of conditioning. procedure will stop
if p > pthr OR abs(z)<zthr OR maxhits hits have been found.}
}
\value{
list of successive significance fine mapped SNPs, named by
    the SNPs
}
\description{
Find independent significant signals in a GWAS summary statistic dataset
}
\author{
Chris Wallace
}
