% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloccc.R
\name{lbf.h3.multinom}
\alias{lbf.h3.multinom}
\title{Approximate LBF under H3, using multinomial approximation}
\usage{
lbf.h3.multinom(bA1, bB1, bA2, bB2, n00, n1, n2, fA, fB, rho, W)
}
\arguments{
\item{bA1}{log OR for SNP A, disease 1 <--- causal for disease 1}

\item{bB1}{log OR for SNP B, disease 1}

\item{bA2}{log OR for SNP A, disease 2}

\item{bB2}{log OR for SNP A, disease 2 <-- causal for disease 2}

\item{n00}{number of shared controls}

\item{n1}{number of disease 1 cases}

\item{n2}{number of disease 1 cases}

\item{fA}{MAF of A in controls}

\item{fB}{MAF of B in controls}

\item{rho}{correlation (r) between A and B in controls}

\item{W}{variance on normal prior for effect sizes at causal variants}
}
\value{
lBF for A, B from the prior involving W to a prior assuming true effect size is 0
}
\description{
Calculate ABF under H3, multinomial
}
\details{
Assumes SNP A is causal for disease 1, and SNP B is causal for disease 2
}
\author{
Chris Wallace
}
