\name{makebinmod.3t}
\alias{makebinmod.3t}
\title{makebinmod}
\usage{
  makebinmod.3t(p)
}
\arguments{
  \item{p}{the number of SNPs}
}
\value{
  a numeric matrix giving the models
}
\description{
  Internal function, makebinmod.3t
}
\details{
  This function takes in a line from the model matrix in
  the 3 trait case and computes which of the categories it
  corresponds to
}
\author{
  Mary Fortune
}

