\name{makebinmod}
\alias{makebinmod}
\title{makebinmod}
\usage{
  makebinmod(p, m)
}
\arguments{
  \item{p}{the number of SNPs}

  \item{m}{the maximum number of causitive snps for each
  trait}
}
\value{
  a numeric matrix giving the models
}
\description{
  Internal function, makebinmod
}
\details{
  This function takes in a line from the model matrix and
  computes which of the categories it corresponds to
}
\author{
  Mary Fortune
}

