% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{map_cond}
\alias{map_cond}
\title{find the next most significant SNP, conditioning on a list
    of sigsnps}
\usage{
map_cond(D, LD, N, YY, sigsnps = NULL)
}
\arguments{
\item{D}{dataset in standard coloc format}

\item{LD}{named matrix of r}

\item{N}{sample size}

\item{YY}{sum(y^2)}

\item{sigsnps}{names of snps to mask}
}
\value{
named numeric - Z score named by snp
}
\description{
Internal helper function for finemap.signals
}
\author{
Chris Wallace
}
