% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R, R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,coloc,missing-method}
\alias{plot,colocBayes,missing-method}
\alias{plot,colocPCs,missing-method}
\alias{plot.coloc_abf}
\title{Plotting functions for the coloc package}
\usage{
plot(x, y, ...)

\S4method{plot}{coloc,missing}(x, y, ...)

\S4method{plot}{colocBayes,missing}(x, y, ...)

\S4method{plot}{colocPCs,missing}(x)

plot.coloc_abf(x, ...)
}
\arguments{
\item{x}{coloc_abf object to be plotted}

\item{y}{ignored}

\item{...}{other arguments}
}
\value{
no return value

ggplot object
}
\description{
You can plot objects of class \code{coloc}, \code{colocBayes} and \code{colocABF}

plot a coloc_abf object
}
\author{
Chris Wallace
}
