% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_datasets}
\alias{plot_datasets}
\title{plot a pair of coloc datasets, highlighting the snps that are common between them}
\usage{
plot_datasets(d1, d2, color = "dodgerblue2")
}
\arguments{
\item{d1}{a coloc dataset}

\item{d2}{another coloc dataset}

\item{color}{optional, specify the colour to use for highlighting}
}
\value{
plots a base graphics figure with two panels
}
\description{
plot a pair of coloc datasets, highlighting the snps that are common between them
}
\author{
Chris Wallace
}
