% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_extended_dataset}
\alias{plot_extended_dataset}
\title{Draw extended plot of summary statistics for two coloc datasets}
\usage{
plot_extended_dataset(
  d,
  x,
  first_highlight_list = NULL,
  second_highlight_list = NULL,
  first_trait = "Trait 1",
  second_trait = "Trait 2",
  snp_label = "snp",
  ens_db = "EnsDb.Hsapiens.v86"
)
}
\arguments{
\item{d}{a coloc dataset}

\item{x}{object of class \code{coloc_abf} returned by coloc.abf()}

\item{first_highlight_list}{character vector of snps to highlight in the first dataset}

\item{second_highlight_list}{character vector of snps to highlight in the second dataset}

\item{first_trait}{label for the first trait}

\item{second_trait}{label for the second trait}

\item{snp_label}{label for the snp column}

\item{ens_db}{character string specifying Ensembl database package from which to get gene positions}
}
\value{
a gtable object
}
\description{
Draw Manhattan-style locus plots for p-values in each dataset, gene annotations,
a scatter plot of z-scores, and a table of coloc summary statistics.
}
\details{
Plot a pair of coloc datasets in an extended format

This function expects that the first two elements of the coloc
dataset list d contain summary statistics. Columns labelled 'chromosome',
'position', and 'p' (p-values) are expected in each dataset. The library
containing the Ensembl database specified by ens_db must be loaded prior to
use (see the example). EnsDb.Hsapiens.v86 is the default database (GRCh38/hg19);
for GRCh37/hg19, use EnsDb.Hsapiens.v75.
}
\examples{
\dontrun{
library(coloc)
library(EnsDb.Hsapiens.v86)
plot(plot_extended_dataset(list(first_dataset, second_dataset), coloc,
first_highlight_list = c("rs123", "rs456"), 
second_highlight_list = c("rs789", "rs1011"), 
ens_db = "EnsDb.Hsapiens.v86"))
}
}
\author{
Tom Willis
}
