% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocqq.R
\name{qq.onesnp}
\alias{qq.onesnp}
\title{qq.onesnp}
\usage{
qq.onesnp(b1, b2, vb1, vb2, fx, v1, v2, eta, N)
}
\arguments{
\item{b1}{coef for snp X, trait 1}

\item{b2}{coef for snp X, trait 2}

\item{vb1}{var(coef) for snp X, trait 1}

\item{vb2}{var(coef) for snp X, trait 2}

\item{fx}{MAF of X}

\item{v1}{variance of trait 1}

\item{v2}{variance of trait 2}

\item{eta}{cor(trait1, trait2)}

\item{N}{number of samples}
}
\value{
var(b1,b2)
}
\description{
reconstruct joint model ioutput for single snp. only vb1b2 really needed
}
\details{
The SNP is X, the traits are 1 and 2

note in this case, the coefficients, and their variances, are
equal in single or joint models.  it is just the var(b1,b2) term
we need.

This provides cor(b1,b2), and must be multiplied by
sqrt(v(b1)*v(b2)) to get var(b1,b2)
}
\author{
Chris Wallace
}
