% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocqq.R
\name{qq.twosnp}
\alias{qq.twosnp}
\title{qq.twosnp}
\usage{
qq.twosnp(b1, b2, g1, g2, vb1, vb2, vg1, vg2, fx, fz, rho, v1, v2, eta, N)
}
\arguments{
\item{b1}{coef for snp X, trait 1}

\item{b2}{coef for snp X, trait 2}

\item{g1}{coef for snp Z, trait 1}

\item{g2}{coef for snp Z, trait 2}

\item{vb1}{var(coef) for snp X, trait 1}

\item{vb2}{var(coef) for snp X, trait 2}

\item{vg1}{var(coef) for snp Z, trait 1}

\item{vg2}{var( coef ) for snp Z, trait 2}

\item{fx}{MAF of X}

\item{fz}{MAF of Z}

\item{rho}{cor(X,Z)}

\item{v1}{variance of trait 1}

\item{v2}{variance of trait 2}

\item{eta}{cor(Y1,Y2)}

\item{N}{number of samples}
}
\value{
list of coefficients and their (co-)variances that would
    be expected from a joint model
}
\description{
reconstruct joint model ioutput for two snp models
}
\details{
The SNPs are X and Z, the traits are 1 and 2
}
\author{
Chris Wallace
}
