% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{subset_dataset}
\alias{subset_dataset}
\title{subset_dataset}
\usage{
subset_dataset(dataset, index)
}
\arguments{
\item{dataset}{coloc dataset}

\item{index}{vector of indices of snps to KEEP}
}
\value{
a copy of dataset, with only the data relating to snps in index remaining
}
\description{
Subset a coloc dataset
}
\author{
Chris Wallace
}
