% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocqq.R
\name{tbf2}
\alias{tbf2}
\title{Bayes factor for one snp model with t distribution}
\usage{
tbf2(b1, b2, v1, v2, rho, w1 = 0.04, w2 = 0.02, N)
}
\arguments{
\item{b1}{coefficient snp A trait 1}

\item{b2}{coefficient snp A trait 2}

\item{v1}{var(b1)}

\item{v2}{var(b2)}

\item{rho}{cor(b1,b2)}

\item{w1}{prior variance for trait 1}

\item{w2}{prior variance for trait 2}

\item{N}{number of samples}
}
\value{
log ABF
}
\description{
Calculate the relative likelihood of a vector of regression
coefficients given their var-covariance matrix under a prior that
the true effect is 0 or is unknown, sampled from a normal with
known variance (W)
}
\details{
This is for single snp models. Assume traits are 1 and 2
}
\author{
Chris Wallace
}
