% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocqq.R
\name{tbf4}
\alias{tbf4}
\title{Bayes factor for two snp model with t distribution}
\usage{
tbf4(bA1, bB1, bA2, bB2, vA1, vA2, vA12, vB1, vB2, vB12, vAB1, vAB2, vAB12, W1,
  W2, N)
}
\arguments{
\item{bA1}{coefficient snp A trait 1}

\item{bB1}{coefficient snp B trait 1}

\item{bA2}{coefficient snp A trait 2}

\item{bB2}{coefficient snp B trait 2}

\item{vA1}{var(bA1)}

\item{vA2}{var(bA2)}

\item{vA12}{var(bA1,bA2)}

\item{vB1}{var(bB1)}

\item{vB2}{var(bB2)}

\item{vB12}{var(bB1,bB2)}

\item{vAB1}{var(bA1,bB1)}

\item{vAB2}{var(bA2,bB2)}

\item{vAB12}{var(bA1,bB2) = var(bB1,bA2)}

\item{W1}{prior variance for trait 1}

\item{W2}{prior variance for trait 1}

\item{N}{number of samples}
}
\value{
log ABF
}
\description{
Calculate the relative likelihood of a vector of regression
coefficients given their var-covariance matrix under a prior that
the true effect is 0 or is unknown, sampled from a normal with
known variance (W)
}
\details{
Assume SNPs are A and B, traits are 1 and 2
}
\author{
Chris Wallace
}
