\name{Var.data.cc}
\alias{Var.data.cc}
\title{Var.data}
\usage{
  Var.data.cc(f, N, s)
}
\arguments{
  \item{s}{???}

  \item{f}{minor allele freq}

  \item{N}{sample number}
}
\value{
  variance of MLE beta
}
\description{
  variance of MLE of beta for case-control
}
\details{
  Internal function
}
\author{
  Claudia Giambartolomei
}

