\name{whichcat.3t}
\alias{whichcat.3t}
\title{whichcat}
\usage{
  whichcat.3t(line)
}
\arguments{
  \item{line}{numeric vector}
}
\value{
  a number corresponding to the category
}
\description{
  Internal function, whichcat.3t
}
\details{
  This function takes in a line from the model matrix and
  computes which of the categories it corresponds to
}
\author{
  Mary Fortune
}

