% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{bf}
\alias{bf}
\alias{bf,colocBayes-method}
\title{Bayes factors to compare specific values of eta}
\usage{
bf(object)

\S4method{bf}{colocBayes}(object)
}
\arguments{
\item{object}{of class \code{colocBayes}}
}
\value{
a matrix of Bayes factors
}
\description{
Summarise the evidence for/against specific values or ranges of eta using bayes factors
}
\details{
Only available for \code{colocBayes} objects, and you need to specify the specific values of interest using the \code{bayes.factor} argument when doing the proportional coloc analysis
}
\author{
Chris Wallace
}
\keyword{methods}
