% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_alignment}
\alias{check_alignment}
\alias{check.alignment}
\title{check alignment}
\usage{
check_alignment(D, thr = 0.2)

check.alignment(...)
}
\arguments{
\item{D}{a coloc dataset}

\item{thr}{plot SNP pairs in absolute LD > thr}

\item{...}{arguments passed to check_alignment()}
}
\value{
a plot for a visual check that alleles in your data are aligned the same way for the beta vector and the LD matrix
}
\description{
check alignment between beta and LD
}
\author{
Chris Wallace
}
