% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dataset}
\alias{check_dataset}
\alias{check.dataset}
\title{check_dataset}
\usage{
check_dataset(d, suffix = "", req = c("snp"), warn.minp = 1e-06)

check.dataset(...)
}
\arguments{
\item{d}{dataset to check}

\item{suffix}{string to identify which dataset (1 or 2)}

\item{req}{names of elements that must be present}

\item{warn.minp}{print warning if no p value < warn.minp}

\item{...}{arguments passed to check_dataset()}
}
\value{
NULL if no errors found
}
\description{
Check coloc dataset inputs for errors
}
\details{
Coloc is flexible, requiring perhaps only p values, or z scores, or effect
estimates and standard errors, but with this flexibility, also comes
difficulties describing exactly the combinations of items required.
\describe{

\item{pvalues}{P-values for each SNP in dataset 1}

\item{N}{Number of samples in dataset 1}

\item{MAF}{minor allele frequency of the variants}

\item{beta}{regression coefficient for each SNP from dataset 1}

\item{varbeta}{variance of beta}

\item{type}{the type of data in dataset 1 - either "quant" or "cc" to denote quantitative or case-control}

\item{s}{for a case control dataset, the proportion of samples in dataset 1 that are cases}

\item{sdY}{for a quantitative trait, the population standard deviation of the trait.  if not given, it can be estimated from the vectors of varbeta and MAF}

\item{snp}{a character vector of snp ids, optional. If present, it will be used to merge dataset1 and dataset2.  Otherwise, the function assumes dataset1 and dataset2 contain results for the same SNPs in the same order.}

}

Some of these items may be missing, but you must always give {\code{type}}.

Then scalars describing the samples used:
\describe{
\item{}{\code{N}}
\item{if \code{type}=="cc"}{\code{s}}
\item{if \code{type}=="quant" and \code{sdY} known}{\code{sdY}}
}
If \code{sdY} is unknown, it will be approximated, and this will require
\describe{
\item{}{\code{beta}, \code{varbeta}, \code{N}, \code{MAF}}
}

Then, if not already covered above, the summary statistics describing the results
\describe{
\item{preferably}{\code{beta}, \code{varbeta}}
\item{alternatively}{\code{pvalues}, \code{MAF}}
}

\code{check_dataset} call stop() unless a series of expectations on dataset
input format are met

This is a helper function for use by other coloc functions, but
you can use it directly to check the format of a dataset to be
supplied to coloc.abf(), coloc.signals(), finemap.abf(), or
finemap.signals().
}
\author{
Chris Wallace
}
