% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc-package.R
\docType{package}
\name{coloc-package}
\alias{coloc-package}
\title{Colocalisation tests of two genetic traits}
\description{
Performs the colocalisation tests described in Plagnol et al (2009) and
Wallace et al (2020) and draws some plots.
}
\author{
Chris Wallace \href{mailto:cew54@cam.ac.uk}{cew54@cam.ac.uk}
}
\keyword{package}
