\docType{package}
\name{coloc-package}
\alias{coloc-package}
\title{Colocalisation tests of two genetic traits}
\description{
  Performs the colocalisation tests described in Plagnol et
  al (2009) and Wallace et al (in preparation) and draws
  some plots.
}
\details{
  \code{coloc.test()} tests for colocalisation and returns
  an object of class \code{coloc}.
}
\author{
  Chris Wallace <chris.wallace@cimr.cam.ac.uk>
}
\references{
  Plagnol et al (2009).  Statistical independence of the
  colocalized association signals for type 1 diabetes and
  RPS26 gene expression on chromosome 12q13. Biostatistics
  10:327-34.

  \url{http://www.ncbi.nlm.nih.gov/pubmed/19039033}

  Wallace et al (in preparation).
}
\keyword{package}

