% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc-package.R
\docType{package}
\name{coloc-package}
\alias{coloc-package}
\title{Colocalisation tests of two genetic traits}
\description{
Performs the colocalisation tests described in Plagnol et al (2009) and
Wallace et al (in preparation) and draws some plots.
}
\details{
\code{coloc.test()} tests for colocalisation and returns an object of class
\code{coloc}.
}
\references{
Plagnol et al (2009).  Statistical independence of the
colocalized association signals for type 1 diabetes and RPS26 gene expression
on chromosome 12q13. Biostatistics 10:327-34.

\url{http://www.ncbi.nlm.nih.gov/pubmed/19039033}

Wallace et al (2013).  Statistical Testing of Shared Genetic Control
for Potentially Related Traits. Genetic Epidemiology 37:802-813.

\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4158901/}

Giambartolomei et al (2014).  Bayesian Test for Colocalisation between Pairs of Genetic Association Studies Using Summary Statistics.  PLOS Genet e1004383.

\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4022491/}
}
\author{
Chris Wallace <chris.wallace@cimr.cam.ac.uk>
}
\keyword{package}
