% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{coloc.signals}
\alias{coloc.signals}
\title{Coloc with multiple signals per trait}
\usage{
coloc.signals(dataset1, dataset2, MAF = NULL, LD = NULL, method = "",
  p1 = 1e-04, p2 = 1e-04, p12 = NULL, maxhits = 3, r2thr = 0.01,
  pthr = 1e-06)
}
\arguments{
\item{dataset1}{a list with the following elements
\describe{

  \item{pvalues}{P-values for each SNP in dataset 1}

  \item{N}{Number of samples in dataset 1}

  \item{MAF}{minor allele frequency of the variants}

\item{beta}{regression coefficient for each SNP from dataset 1}

\item{varbeta}{variance of beta}

\item{type}{the type of data in dataset 1 - either "quant" or "cc" to denote quantitative or case-control}

\item{s}{for a case control dataset, the proportion of samples in dataset 1 that are cases}

 \item{sdY}{for a quantitative trait, the population standard deviation of the trait.  if not given, it can be estimated from the vectors of varbeta and MAF}

\item{snp}{a character vector of snp ids, optional. If present, it will be used to merge dataset1 and dataset2.  Otherwise, the function assumes dataset1 and dataset2 contain results for the same SNPs in the same order.}

}

Some of these items may be missing, but you must give
\itemize{
\item{always}{\code{type}}
\item{if \code{type}=="cc"}{\code{s}}
\item{if \code{type}=="quant" and \code{sdY} known}{\code{sdY}}
\item{if \code{type}=="quant" and \code{sdY} unknown}{\code{beta}, \code{varbeta}, \code{N}, \code{MAF}}
and then either
\item{}{\code{pvalues}, \code{MAF}}
\item{}{\code{beta}, \code{varbeta}}
}}

\item{dataset2}{as above, for dataset 2}

\item{MAF}{Common minor allele frequency vector to be used for both dataset1 and dataset2, a shorthand for supplying the same vector as parts of both datasets}

\item{LD}{required if method="cond". matrix of genotype
*correlation* (ie r, not r^2) between SNPs. If dataset1 and
dataset2 may have different LD, you can instead add LD=LD1 to
the list of dataset1 and a different LD matrix for dataset2}

\item{method}{default "" means do no conditioning, should return
similar to coloc.abf.  if method="cond", then use conditioning
to coloc multiple signals.  if method="mask", use masking to
coloc multiple signals. if different datasets need different
methods (eg LD is only available for one of them) you can set
method on a per-dataset basis by adding method="..." to the
list for that dataset.}

\item{p1}{prior probability a SNP is associated with trait 1, default 1e-4}

\item{p2}{prior probability a SNP is associated with trait 2, default 1e-4}

\item{p12}{prior probability a SNP is associated with both traits, default 1e-5}

\item{maxhits}{maximum number of levels to condition/mask}

\item{r2thr}{if masking, the threshold on r2 should be used to
call two signals independent.  our experience is that this
needs to be set low to avoid double calling the same strong
signal.}

\item{pthr}{if masking or conditioning, what p value threshold to
call a secondary hit "significant"}
}
\value{
data.table of coloc results, one row per pair of lead snps
    detected in each dataset
}
\description{
New coloc function, builds on coloc.abf() by allowing for multiple
independent causal variants per trait through conditioning or
masking.
}
\author{
Chris Wallace
}
