% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{coloc.susie}
\alias{coloc.susie}
\title{run coloc using susie to detect separate signals}
\usage{
coloc.susie(dataset1, dataset2, ...)
}
\arguments{
\item{dataset1}{\emph{either} a coloc-style input dataset, or the result of
running runsusie on such a dataset}

\item{dataset2}{\emph{either} a coloc-style input dataset, or the result of
running runusie on such a dataset}

\item{...}{other arguments passed to \link{coloc.bf_bf}}
}
\value{
coloc.signals style result
}
\description{
colocalisation with multiple causal variants via SUSIE
}
\details{
.. content for \details{} ..
}
\author{
Chris Wallace
}
