% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claudia.R
\name{credible.sets}
\alias{credible.sets}
\title{credible.sets}
\usage{
credible.sets(dataset, credible.size = 0.95)
}
\arguments{
\item{dataset}{data.frame output of \code{finemap.abf()}}

\item{credible.size}{threshold of the credible set (Default: 0.95)}
}
\value{
SNP ids of the credible set
}
\description{
Get credible sets from finemapping results
}
\author{
Guillermo Reales, Chris Wallace
}
