% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{est_cond}
\alias{est_cond}
\title{generate conditional summary stats}
\usage{
est_cond(x, LD, YY, sigsnps, xtx = NULL)
}
\arguments{
\item{x}{coloc dataset}

\item{LD}{named matrix of r}

\item{YY}{sum((Y-Ybar)^2)}

\item{sigsnps}{names of snps to jointly condition on}

\item{cc}{TRUE if case control}
}
\value{
data.table giving snp, beta and varbeta on remaining snps
after conditioning
}
\description{
Internal helper function for est_all_cond
}
\author{
Chris Wallace
}
