% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{est_cond}
\alias{est_cond}
\title{generate conditional summary stats}
\usage{
est_cond(x, LD, N, YY, sigsnps, cc = FALSE)
}
\arguments{
\item{x}{data.table created from standard coloc dataset}

\item{LD}{named matrix of r}

\item{N}{sample size}

\item{YY}{sum(y^2)}

\item{sigsnps}{names of snps to mask}

\item{cc}{TRUE if case control}
}
\value{
data.table giving snp, beta and varbeta on remaining snps after conditioning
}
\description{
Internal helper function for map_cond
}
\author{
Chris Wallace
}
