% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{estgeno.1.ctl}
\alias{estgeno.1.ctl}
\alias{estgeno.1.cse}
\title{estgeno1}
\usage{
estgeno.1.ctl(f)

estgeno.1.cse(G0, b)
}
\arguments{
\item{f}{MAF}

\item{G0}{single snp frequency in controls (vector of length 3) -
obtained from estgeno.1.ctl}

\item{b}{log odds ratio}
}
\value{
relative frequency of genotypes 0, 1, 2
}
\description{
Estimate single snp frequency distibutions
}
\seealso{
estgeno2
}
\author{
Chris Wallace
}
