% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{eta}
\alias{eta}
\alias{theta}
\alias{summary}
\alias{ci}
\alias{eta,coloc-method}
\alias{theta,coloc-method}
\alias{summary,coloc-method}
\alias{ci,coloc-method}
\alias{eta,colocBayes-method}
\alias{theta,colocBayes-method}
\alias{summary,colocBayes-method}
\alias{ci,colocBayes-method}
\title{Methods to extract information from a \code{coloc} or \code{colocBayes}
object}
\usage{
eta(object)
}
\arguments{
\item{object}{Object returned by \code{coloc.test()} or \code{coloc.bma()} functions.}
}
\description{
Extract information from a \code{coloc} object.
}
\details{
\code{eta()} returns eta.hat, the maximum likelihood value of eta.

\code{theta()} returns theta.hat, the maximum likelihood value of eta.

\code{summary()} returns a summary, giving eta, chisquare statistic, number of SNPs/PCs, p value and, if a \code{colocBayes} object, the ppp.value

\code{ci()} returns the credible interval, or \code{NA} if not calculated.
}
\seealso{
\code{\link{coloc.test}}, \code{\link{pcs.prepare}}
}
\author{
Chris Wallace.
}
\keyword{methods}
