% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot.coloc_abf}
\alias{plot.coloc_abf}
\title{plot a coloc_abf object}
\usage{
\method{plot}{coloc_abf}(x, ...)
}
\arguments{
\item{x}{coloc_abf object to be plotted}

\item{...}{other arguments}
}
\value{
ggplot object
}
\description{
plot a coloc_abf object
}
\author{
Chris Wallace
}
