% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dataset}
\alias{plot_dataset}
\title{plot a coloc dataset}
\usage{
plot_dataset(d, ...)
}
\arguments{
\item{d}{a coloc dataset}

\item{...}{other arguments passed to the base graphics plot() function}
}
\description{
Plot a coloc structured dataset
}
\author{
Chris Wallace
}
