% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_extended_dataset}
\alias{plot_extended_dataset}
\title{Draw extended plot of summary statistics for two coloc datasets}
\usage{
plot_extended_dataset(
  dataset1,
  dataset2,
  x,
  first_highlight_list = NULL,
  second_highlight_list = NULL,
  first_index_snp = NULL,
  second_index_snp = NULL,
  first_trait = "Trait 1",
  second_trait = "Trait 2",
  snp_label = "snp",
  ld_label = NULL,
  show_ld = FALSE,
  locus_plot_ylim = NULL,
  ens_db = "EnsDb.Hsapiens.v86"
)
}
\arguments{
\item{dataset1}{a coloc dataset}

\item{dataset2}{a coloc dataset}

\item{x}{object of class \code{coloc_abf} returned by \code{coloc.abf()}}

\item{first_highlight_list}{character vector of snps to highlight in the first dataset;
\code{'index'} can be passed to highlight the most significant SNP}

\item{second_highlight_list}{character vector of snps to highlight in the second dataset;
\code{'index'} can be passed to highlight the most significant SNP}

\item{first_index_snp}{snp to designate as index snp in the first dataset
for the purpose of LD visualisation}

\item{second_index_snp}{snp to designate as index snp in the second dataset
for the purpose of LD visualisation}

\item{first_trait}{label for the first trait}

\item{second_trait}{label for the second trait}

\item{snp_label}{label for the snp column}

\item{ld_label}{label for the LD column}

\item{show_ld}{logical, whether to show LD in the locus plots}

\item{locus_plot_ylim}{numeric vector of length 2 specifying the y-axis
limits for the locus plots on the -log10 scale}

\item{ens_db}{character string specifying Ensembl database package from which to get gene positions. Current (at time of writing this documentation!) sensible values are  "EnsDb.Hsapiens.v86" for build 38 and "EnsDb.Hsapiens.v75" for build 37.}
}
\value{
a gtable object
}
\description{
Draw Manhattan-style locus plots for p-values in each dataset, gene annotations,
a scatter plot of z-scores, and a table of coloc summary statistics.
}
\details{
This function expects that the first two elements of the coloc
dataset list d contain summary statistics. Columns labelled 'chromosome' and
'position' are expected in each dataset.

The packages \code{locuszoomr} and one of \code{EnsDb.Hsapiens.v86} or \code{EnsDb.Hsapiens.v75} are required
for this function. One of the \code{EnsDb.Hsapiens} libraries containing the Ensembl database
specified by ens_db must be loaded prior to  use (see the example). EnsDb.Hsapiens.v86
is the default database (GRCh38/hg19); for GRCh37/hg19, use EnsDb.Hsapiens.v75.

If \code{beta} and \code{varbeta} are absent, -log10(p) values will be plotted
in lieu of z-scores.
}
\examples{
\dontrun{
library(coloc)
library(EnsDb.Hsapiens.v86)
plot(plot_extended_dataset(list(first_dataset, second_dataset), coloc,
first_highlight_list = c("rs123", "rs456"), 
second_highlight_list = c("rs789", "rs1011"), 
ens_db = "EnsDb.Hsapiens.v86"))
}
}
\author{
Tom Willis

Chris Wallace
}
