% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{print.coloc_abf}
\alias{print.coloc_abf}
\title{print.coloc_abf}
\usage{
\method{print}{coloc_abf}(x, trait1 = "trait 1", trait2 = "trait 2")
}
\arguments{
\item{x}{object of class \code{coloc_abf} returned by coloc.abf()
or coloc.signals()}

\item{trait1}{name of trait 1 (optional)}

\item{trait2}{name of trait 2 (optional)}
}
\value{
x, invisibly
}
\description{
Print summary of a coloc.abf run
}
\author{
Chris Wallace
}
