% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{runsusie}
\alias{runsusie}
\title{Run susie on a single coloc-structured dataset}
\usage{
runsusie(
  d,
  suffix = 1,
  nref = 503,
  p = 1e-04,
  trimz = NULL,
  L = 10,
  r2.prune = NULL,
  s_init = NULL
)
}
\arguments{
\item{d}{coloc dataset, must include LD (signed correlation matrix)}

\item{suffix}{suffix label that will be printed with any error messages}

\item{nref}{number of samples in the dataset used to estimate the LD matrix in d}

\item{p}{prior probability a snp is causal (equivalent to p1 or p2 in coloc.abf)}

\item{trimz}{used to trim datasets for development purposes}

\item{L}{maximum number of signals to consider, passed to susie}

\item{r2.prune}{sometimes SuSiE can return multiple signals in high LD. if you set r2.prune to a value between 0 and 1, sets with index SNPs with LD > r2.prune}

\item{s_init}{used internally to extend runs that haven't converged. don't use.}
}
\value{
results of a susie_rss run, with some added dimnames
}
\description{
run susie_rss storing some additional information for coloc
}
\examples{
library(coloc)
data(coloc_test_data)
result=runsusie(coloc_test_data$D1)
}
\author{
Chris Wallace
}
