% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{runsusie}
\alias{runsusie}
\title{Run susie on a single coloc-structured dataset}
\usage{
runsusie(d, suffix = 1, nref = 503)
}
\arguments{
\item{d}{dataset}

\item{suffix}{suffix label that will be printed with any error messages}
}
\value{
results of a susie_rss run, with some added dimnames
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Chris Wallace
}
